validation for GET and DELETE

+++ app/Http/Controllers/GenericResourceController.php
@@ -32,6 +32,10 @@ class GenericResourceController extends Controller
             return $this->jsonError(['Model not found.'], 404);
         }
 
+        if ($this->validateInputsForResource($request->all(), $request->route('resource')) === false) {
+            return $this->jsonError(['Insufficient permissions.'], 403);
+        }
+
         return $this->jsonSuccess($model);
     }
 
@@ -86,6 +90,10 @@ class GenericResourceController extends Controller
 
         if (!$model instanceof GenericModel) {
             return $this->jsonError(['Model not found.'], 404);
+        }
+
+        if ($this->validateInputsForResource($request->all(), $request->route('resource')) === false) {
+            return $this->jsonError(['Insufficient permissions.'], 403);
         }
 
         if ($model->delete()) {
